Make vim-gcov-marker customizable (fixes #4)
Includes:
choose whether or not we open the location list
customize markers (the ones used currently are Unicode and they might not be displayed correctly in old vim terminal setup)
put SetCov in autoload, this way it is lazy loaded.
Reflect changes in theREADME and the docs

Files were renamed using underscore instead of hypher for autoloading purpose only.